Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.15.2
motor==3.3.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
aiofiles==23.2.1
loguru==0.7.2
python-multipart==0.0.6
beautifulsoup4==4.12.2
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.6.1
aiohttp==3.9.1 